game info now opens in modal
