--- pages/game/[id].js
@@ -1,7 +1,0 @@
-import Game from '../../src/components/Game'
-
-const GamePage = () => {
-	return <Game />
-}
-
-export default GamePage
