+++ src/components/Backlog.js
@@ -9,14 +9,10 @@ import { Add } from '@mui/icons-material'
 import { useTheme } from '@mui/styles'
 
 import { useAppContext } from '/src/AppContext'
+import Layout from '@components/layout'
 
-import {
-	AddGameModal,
-	AppBar,
-	NavTabs,
-	GameCards,
-	LoadingOverlay,
-} from './components'
+import { AddGameModal, GameCards, LoadingOverlay } from './components'
+import GameModal from './GameModal'
 
 const Backlog = () => {
 	const { data, status } = useSession()
@@ -24,13 +20,15 @@ const Backlog = () => {
 	const isMobile = useMediaQuery(theme.breakpoints.down('md'))
 
 	const { query } = useRouter()
-	const { games, handleApi, loading } = useAppContext()
+	const { games, handleApi, loading, filter } = useAppContext()
 
-	const [filter, setFilter] = useState(query.tab || 'not_started')
 	const [title, setTitle] = useState('Backlog')
 	const [filteredGames, setFilteredGames] = useState()
 	const [openModal, setOpenModal] = useState(false)
 	const [viewMode, setViewMode] = useState('grid')
+
+	const [modalId, setModalId] = useState()
+	const [gameModalOpen, setGameModalOpen] = useState(false)
 
 	useEffect(() => {
 		if (games) {
@@ -65,8 +63,13 @@ const Backlog = () => {
 		}
 	}, [filter])
 
+	const closeGameModal = () => {
+		setGameModalOpen(false)
+		setModalId(null)
+	}
+
 	return (
-		<>
+		<Layout title={title}>
 			{!status === 'authenticated' ? null : (
 				<>
 					{data?.user.name ? (
@@ -81,10 +84,8 @@ const Backlog = () => {
 						sx={{
 							width: '100%',
 							paddingBottom: { xs: 22, md: 9 },
-							paddingLeft: { xs: 0, md: 31 },
 						}}
 					>
-						<AppBar title={title} />
 						<Box
 							sx={{
 								margin: 2,
@@ -112,15 +113,20 @@ const Backlog = () => {
 						<GameCards
 							games={filteredGames}
 							loading={loading}
-							handleApi={handleApi}
+							setModalId={setModalId}
+							setModalOpen={setGameModalOpen}
 						/>
 					</Box>
 					<AddGameModal openModal={openModal} setOpenModal={setOpenModal} />
-					<NavTabs setFilter={setFilter} />
+					<GameModal
+						id={modalId}
+						open={gameModalOpen}
+						modalClose={closeGameModal}
+					/>
 				</>
 			)}
 			<LoadingOverlay loading={loading} />
-		</>
+		</Layout>
 	)
 }
