+++ src/components/GameModal/components/ConfirmModal.js
