+++ src/components/GameModal/components/StatusButton.js
