+++ src/components/GameModal/components/index.js
