+++ src/components/Layout/components/index.js
@@ -1,0 +1,4 @@
+import AppBar from './AppBar'
+import NavTabs from './NavTabs'
+
+export { AppBar, NavTabs }
