+++ src/components/components/index.js
@@ -1,15 +1,6 @@
 import AddGameModal from './AddGameModal'
-import AppBar from './AppBar'
-import NavTabs from './NavTabs'
 import GameCards from './GameCards'
 import LoadingOverlay from './Loading'
 import PlatformIcon from './PlatformIcon'
 
-export {
-	AddGameModal,
-	AppBar,
-	NavTabs,
-	GameCards,
-	LoadingOverlay,
-	PlatformIcon,
-}
+export { AddGameModal, GameCards, LoadingOverlay, PlatformIcon }
